// src/types/character.ts
export interface Character {
    id: number;
    name: string;
    image: string;
    nationality: string;
    ethnicity: string;
    gender: string;
    eyeColor: string;
    hairColor: string;
    skinColor: string;
    allies: string[];
    enemies: string[];
    weaponOfChoice: string[];
    fightingStyles: string[];
    firstAppearance: string;
    voiceActors: string[];
}